+++ app/guid-node/forks/route.ts
@@ -1,9 +1,11 @@
 import { service } from '@ember-decorators/service';
 import Route from '@ember/routing/route';
 
+import featureFlagRoute from 'ember-osf-web/decorators/feature-flag-route';
 import GuidNodeForksController from 'ember-osf-web/guid-node/forks/controller';
 import Ready from 'ember-osf-web/services/ready';
 
+@featureFlagRoute()
 export default class GuidNodeForks extends Route {
     @service ready!: Ready;